+++ 5.Ultrasound_Serial/Timer0.h
@@ -6,6 +6,6 @@
 void Timer0_Init();
 void Timer0_Start();
 void Timer0_Stop();
-unsigned int Timer0_CurrentCount();
+unsigned short Timer0_CurrentCount();
 
 #endif
